+++ coffee/app/models/Schedule.coffee
@@ -6,7 +6,8 @@ db.execute "CREATE TABLE IF NOT EXISTS SCHEDULEDB (ID INTEGER PRIMARY KEY, TITLE
 
 class Schedule
 #   itype -1, history 0, timer 1, favorite 2,  folder 3 
-  constructor: (@title, @active = 0, @updated = -1, @id = null, @date = (new Date()).getTime(), @scheme = 'http://www.google.com', @repeat = 0, @options = {}, @isChanged = false) ->    
+  constructor: (@title, @active = 0, @updated = -1, @id = null, @date = (new Date()).getTime(), @scheme = 'http://www.google.com', @repeat = 0, @options = {}) ->
+    Ti.App.Properties.removeProperty 'lastSchedule'
         
   save: () ->
     now = (new Date()).getTime()
@@ -15,11 +16,12 @@ class Schedule
       @id = db.lastInsertRowId
     else
       db.execute "UPDATE SCHEDULEDB SET TITLE = ?,ACTIVE = ? ,DATE = ? ,SCHEME = ? ,REPEAT = ? ,UPDATED = ? ,OPTIONS = ?  WHERE id = ?", @title, @active, @date, @scheme, @repeat, now, JSON.stringify(@options), @id
-    @isChanged = false
+    Ti.App.Properties.setInt 'lastSchedule',  @id
     return this
     
   del: () ->
-    db.execute "DELETE FROM SCHEDULEDB WHERE id = ?", @id
+    if @id isnt null
+      db.execute "DELETE FROM SCHEDULEDB WHERE id = ?", @id
     return null
     
   @findAll: (sql) ->
@@ -45,6 +47,7 @@ class Schedule
     if rows.isValidRow()      
       f = rows.fieldByName
       schedule = new Schedule f('TITLE'), f('ACTIVE'), parseInt(f('UPDATED'), 10), f('ID'), parseInt(f('DATE'), 10), f('SCHEME'), f('REPEAT'), JSON.parse(f('OPTIONS'))
+      Ti.App.Properties.setInt 'lastSchedule',  f('ID')
     rows.close()    
     return schedule    
     
@@ -64,8 +67,10 @@ class Schedule
     Schedule.count "SELECT ID FROM SCHEDULEDB WHERE ACTIVE > 0"
     
   @findById: (id) ->
-    Ti.App.Properties.setInt 'lastSchedule',  id
     Schedule.findOne "SELECT * FROM SCHEDULEDB WHERE ID = #{id}"
+    
+  @findLastUpdated: () ->
+    Schedule.findOne "SELECT * FROM SCHEDULEDB  ORDER BY UPDATED"
     
   @delAll: () ->
     db.execute "DELETE FROM SCHEDULEDB"
